rename to useUserTenantPermissions
