+++ src/hooks/index.js
@@ -1,1 +1,1 @@
-export { default as useUserSelfTenantPermissions } from './useUserSelfTenantPermissions';
+export { default as useUserTenantPermissions } from './useUserTenantPermissions'; // eslint-disable-line import/prefer-default-export
